refactor: global/success 폴더 구조 변경 (#29)
