+++ src/main/java/eightbit/moyeohaeng/global/success/SuccessCode.java
@@ -1,4 +1,4 @@
-package eightbit.moyeohaeng.global.success.common;
+package eightbit.moyeohaeng.global.success;
 
 import org.springframework.http.HttpStatus;
